+++ braintree/authorization_adjustment.py
@@ -4,5 +4,5 @@ from braintree.attribute_getter import AttributeGetter
 class AuthorizationAdjustment(AttributeGetter):
     def __init__(self, attributes):
         AttributeGetter.__init__(self, attributes)
-        if hasattr(self, 'amount') and self.amount is not None:
+        if getattr(self, "amount", None) is not None:
             self.amount = Decimal(self.amount)