+++ braintree/util/http.py
@@ -1,9 +1,6 @@
 import sys
 import requests
-if sys.version_info[0] == 2:
-    from base64 import encodestring as encodebytes
-else:
-    from base64 import encodebytes
+from base64 import encodebytes
 import json
 import braintree
 from braintree import version
@@ -11,17 +8,19 @@ from braintree.environment import Environment
 from braintree.util.xml_util import XmlUtil
 from braintree.exceptions.authentication_error import AuthenticationError
 from braintree.exceptions.authorization_error import AuthorizationError
-from braintree.exceptions.down_for_maintenance_error import DownForMaintenanceError
-from braintree.exceptions.not_found_error import NotFoundError
-from braintree.exceptions.server_error import ServerError
-from braintree.exceptions.too_many_requests_error import TooManyRequestsError
-from braintree.exceptions.upgrade_required_error import UpgradeRequiredError
-from braintree.exceptions.unexpected_error import UnexpectedError
+from braintree.exceptions.gateway_timeout_error import GatewayTimeoutError
 from braintree.exceptions.http.connection_error import ConnectionError
 from braintree.exceptions.http.invalid_response_error import InvalidResponseError
-from braintree.exceptions.http.timeout_error import TimeoutError
 from braintree.exceptions.http.timeout_error import ConnectTimeoutError
 from braintree.exceptions.http.timeout_error import ReadTimeoutError
+from braintree.exceptions.http.timeout_error import TimeoutError
+from braintree.exceptions.not_found_error import NotFoundError
+from braintree.exceptions.request_timeout_error import RequestTimeoutError
+from braintree.exceptions.server_error import ServerError
+from braintree.exceptions.service_unavailable_error import ServiceUnavailableError
+from braintree.exceptions.too_many_requests_error import TooManyRequestsError
+from braintree.exceptions.unexpected_error import UnexpectedError
+from braintree.exceptions.upgrade_required_error import UpgradeRequiredError
 
 class Http(object):
     class ContentType(object):
@@ -41,6 +40,8 @@ class Http(object):
             raise AuthorizationError(message)
         elif status == 404:
             raise NotFoundError()
+        elif status == 408:
+            raise RequestTimeoutError()
         elif status == 426:
             raise UpgradeRequiredError()
         elif status == 429:
@@ -48,7 +49,9 @@ class Http(object):
         elif status == 500:
             raise ServerError()
         elif status == 503:
-            raise DownForMaintenanceError()
+            raise ServiceUnavailableError()
+        elif status == 504:
+            raise GatewayTimeoutError()
         else:
             raise UnexpectedError("Unexpected HTTP_RESPONSE " + str(status))